Corrected a bunch of pedantic warnings

+++ lib/my/src/misc/find_prime_sup.c
@@ -12,8 +12,8 @@ int my_find_prime_sup(int number)
 {
     if (number < 0)
         number = 0;
-    for (int i = number; i < INT_MAX; ++i)
+    for (unsigned i = (unsigned)number; i < INT_MAX; ++i)
         if (my_is_prime(i))
-            return (i);
+            return ((int)i);
     return (0);
 }
